// src/app/models/invoice.ts
export class Invoice {
    refId: number;
    orderNumber: number;
    showroom: string;
    itemName: string;
    grossWeight: number;
    totalStones: number;
    stoneWeightInCarat: number;
    stoneWeightInGrams: number;
    netWeight: number;
    gold92Per: string;
    gold12Per: string;
    deliveryDate: Date;
    makingCharges: number;
    stoneWeight: number;
}